requirements.txt
langchain
langchain-groq
groq
streamlit
langchain-community
python-dotenv

